Feature3/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
requests
groq
python-dotenv